made changed to selection sort
