+++ src/core/main.cpp
@@ -30,7 +30,6 @@ int main(int argc, char* argv[]) {
 
   // std::vector<std::unique_ptr<Shape>> shapes;
   // Utility::Parser::parseShapes(shapes);
-  Test::ints();
 
   return app.exec();
 }
