+++ src/gm/selection-sort.h
@@ -1,18 +1,17 @@
 #pragma once
 
 #include "gm/vector.h"
-#include "models/free-form/circle.h"
 
 template <typename T>
 bool compare_shape_perimeter(
     typename GM::Vector<T>::iterator &it, typename GM::Vector<T>::iterator &minIt) {
-  return (*it).perimeter() < (*minIt).perimeter();
+  return it->perimeter() < minIt->perimeter();
 }
 
 template <typename T>
 bool compare_shape_area(
     typename GM::Vector<T>::iterator &it, typename GM::Vector<T>::iterator &minIt) {
-  return (*it).area() < (*minIt).area();
+  return it->area() < minIt->area();
 }
 
 // sorts by function pointer
@@ -42,7 +41,7 @@ void selection_sort(GM::Vector<T> &vec) {
   for (it itor = vec.begin(); itor != vec.end(); ++itor) {
     it minItor = itor;
     for (it i = itor + 1; i != vec.end(); ++i) {
-      if ((*i).getShapeId() < (*minItor).getShapeId()) {
+      if (i->getShapeId() < minItor->getShapeId()) {
         minItor = i;
       }
     }
@@ -51,17 +50,3 @@ void selection_sort(GM::Vector<T> &vec) {
     }
   }
 }
-
-namespace Test {
-
-template <typename T>
-void print(const GM::Vector<T> &vec) {
-  for (auto itor = vec.begin(); itor != vec.end(); itor++) {
-    std::cout << *itor << " ";
-  }
-  std::cout << '\n';
-}
-
-void sortingShapeByPer();
-void sortingShapeById();
-}  // namespace Test
