+++ src/tests/selection-sort/sort-test.cpp
@@ -1,4 +1,4 @@
-#include "selection-sort.h"
+#include "sort-test.h"
 
 namespace Test {
